+++ Models/BalanceModel.cs
@@ -6,28 +6,28 @@ public class Balance
 {
     public int BalanceId { get; set; }
 
-    public int? Type1 { get; set; } //Checking
-    public int? Type2 { get; set; } //Currency
-    public int? Type3 { get; set; } //Savings
-    public int? Type4 { get; set; } //CDs
-    public int? Type5 { get; set; } //Health and Medical Savings
-    public int? Type6 { get; set; } //Life Insurance
-    public int? Type7 { get; set; } //Annuities
-    public int? Type8 { get; set; } //Investments
-    public int? Type9 { get; set; } //IRAs
-    public int? Type10 { get; set; } //Roth IRA
-    public int? Type11 { get; set; } //Employer Retirement Plans
-    public int? Type12 { get; set; } //Bullion
-    public int? Type13 { get; set; } //Primary Residence
-    public int? Type14 { get; set; } //Secondary Residence
-    public int? Type15 { get; set; } //Real Estate
-    public int? Type16 { get; set; } //Business
-    public int? Type17 { get; set; } //Trust
-    public int? Type18 { get; set; } //Vehicles
-    public int? Type19 { get; set; } //Personal Property
-    public int? Type20 { get; set; } //Credit Cards
-    public int? Type21 { get; set; } //Student Loans
-    public int? Type22 { get; set; } //Other
+    public double? Type1 { get; set; } //Checking
+    public double? Type2 { get; set; } //Currency
+    public double? Type3 { get; set; } //Savings
+    public double? Type4 { get; set; } //CDs
+    public double? Type5 { get; set; } //Health and Medical Savings
+    public double? Type6 { get; set; } //Life Insurance
+    public double? Type7 { get; set; } //Annuities
+    public double? Type8 { get; set; } //Investments
+    public double? Type9 { get; set; } //IRAs
+    public double? Type10 { get; set; } //Roth IRA
+    public double? Type11 { get; set; } //Employer Retirement Plans
+    public double? Type12 { get; set; } //Bullion
+    public double? Type13 { get; set; } //Primary Residence
+    public double? Type14 { get; set; } //Secondary Residence
+    public double? Type15 { get; set; } //Real Estate
+    public double? Type16 { get; set; } //Business
+    public double? Type17 { get; set; } //Trust
+    public double? Type18 { get; set; } //Vehicles
+    public double? Type19 { get; set; } //Personal Property
+    public double? Type20 { get; set; } //Credit Cards
+    public double? Type21 { get; set; } //Student Loans
+    public double? Type22 { get; set; } //Other
 
     public string? Username{get; set;}
